+++ example/lib/main.dart
@@ -23,7 +23,12 @@ class _MyAppState extends State<MyApp> {
     try {
       //LINK CONTAINS APK OF FLUTTER HELLO WORLD FROM FLUTTER SDK EXAMPLES
       OtaUpdate()
-          .execute('https://internal1.4q.sk/flutter_hello_world.apk', destinationFilename: 'flutter_hello_world.apk')
+          .execute(
+        'https://internal1.4q.sk/flutter_hello_world.apk',
+        destinationFilename: 'flutter_hello_world.apk',
+        //FOR NOW ANDROID ONLY - ABILITY TO VALIDATE CHECKSUM OF FILE:
+        sha256checksum: "d6da28451a1e15cf7a75f2c3f151befad3b80ad0bb232ab15c20897e54f21478",
+      )
           .listen(
         (OtaEvent event) {
           setState(() => currentEvent = event);
@@ -45,8 +50,7 @@ class _MyAppState extends State<MyApp> {
           title: const Text('Plugin example app'),
         ),
         body: Center(
-          child: Text(
-              'OTA status: ${currentEvent.status} : ${currentEvent.value} \n'),
+          child: Text('OTA status: ${currentEvent.status} : ${currentEvent.value} \n'),
         ),
       ),
     );